+++ docs/source/_ext/generate_contrib_docs.py
@@ -2,12 +2,12 @@ from pathlib import Path
 import importlib
 import sys
 
-import cofi_espresso as esp
+import espresso
 
 
 # APV removed following as it is never used(?)
 # CONTRIBS = [nm for nm in dir(esp) if not nm.startswith("_") and nm[0].islower()]
-BASE_PATH = esp.__path__[0]
+BASE_PATH = espresso.__path__[0]
 DEST_PATH = Path(__file__).resolve().parent.parent / "user_guide" / "contrib" / "generated"
 ROOT_PATH = Path(__file__).resolve().parent.parent.parent.parent
 _machine_doc_utils_gen_docs = ROOT_PATH / "espresso_machine" / "doc_utils"
@@ -16,7 +16,7 @@ try:
     sys.path.append(str(_machine_doc_utils_gen_docs))
     import gen_docs
 except:
-    from esp._machine import gen_docs           # TODO test this after build.py is finished
+    from espresso._machine.doc_utils import gen_docs           # TODO test this after build.py is finished
 
 
 def setup(app):