+++ espresso_machine/build_package/build.py
@@ -1,10 +1,11 @@
-"""Build the Python package "cofi_espresso"
+"""Build the Python package "espresso"
 
 1. clean "_esp_build/"
 2. "/<meta-data-files>" => "_esp_build/"
 3. "src/" => "_esp_build/src/"
-4. "contrib/" => "_esp_build/src/cofi_espresso/" + "__init__.py" + "list_problems.py"
+4. "contrib/" => "_esp_build/src/espresso/" + "__init__.py" + "list_problems.py"
 5. "_version.py" => "_esp_build/pyproject.toml"
+6. "espresso_machine/" => _esp_build/src/_machine"
 6. `pip install .`
 
 """
@@ -20,7 +21,7 @@ import versioningit
 
 
 # ------------------------ constants ------------------------
-PKG_NAME = "cofi_espresso"
+PKG_NAME = "espresso"
 current_directory = Path(__file__).resolve().parent
 root = current_directory.parent.parent
 ROOT_DIR = str(root)
@@ -29,6 +30,7 @@ PKG_SRC = str(root / "src")
 CONTRIB_SRC = str(root / "contrib")
 VCS_GIT = str(root / ".git")
 DOCS_SRC = str(root / "docs")
+MACHINE_SRC = str(root / "espresso_machine")
 META_FILES = [
     "README.md",
     "pyproject.toml",
@@ -153,7 +155,7 @@ def write_version():
             "distance-dirty": "{base_version}+{distance}.{vcs}{rev}.dirty",
         },
         "write": {
-            "file": "src/cofi_espresso/_version.py"
+            "file": "src/espresso/_version.py"
         }
     }
     version = versioningit.get_version(root, versioningit_config, True)
@@ -173,7 +175,11 @@ def write_version():
     with open(f"{BUILD_DIR}/pyproject.toml", "w") as f:
         f.write(file_content)
 
-# 6
+# 6 move espresso_machine into espresso/_machine
+def move_espresso_machine():
+    move_folder_content(MACHINE_SRC, f"{BUILD_DIR}/src/espresso/_machine")
+
+# 7
 def install_pkg():
     subprocess.call([sys.executable, "-m", "pip", "uninstall", "-y", PKG_NAME])
     return subprocess.call([sys.executable, "-m", "pip", "install", "."], cwd=BUILD_DIR)
@@ -210,7 +216,11 @@ def build():
     write_version()
     print("OK.")
     # 6
-    println_with_emoji("Building Python package: cofi-espresso...", "🗂")
+    println_with_emoji("Moving infrastructure code...", "🗂")
+    move_espresso_machine()
+    print("OK.")
+    # 6
+    println_with_emoji("Building Python package: geo-espresso...", "🗂")
     exit_code = install_pkg()
     if exit_code == 0: 
         println_with_emoji("Espresso installed!", "🍰")